refactor: MultiSplitView Provider와 Notifier 분리
